requirements.txt
flask>=3.0.0
google-genai>=1.0.0
yt-dlp>=2024.1.0